requirements.txt
docling==2.18.0
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.21
langchain-mongodb==0.2.0
langchain-openai==0.2.11
langchain-text-splitters==0.3.2
langgraph==0.2.56
pymongo==4.10.1
python-dotenv==1.0.1
streamlit==1.42.1
pytz==2024.1
tiktoken==0.8.0
Pillow==10.4.0
python-docx==1.1.2
pymupdf==1.24.3
openai==1.63.2
azure-core==1.30.1
azure-ai-formrecognizer==3.3.0
azure-identity==1.16.1